neida

+++ Python - Mundo 1/Exercícios/Ex008.py
@@ -1,5 +1,5 @@
 # Escreva um programa que leia um valor em metros e o axiba convertido em centímetros e milímetros.
 
-n = int(input('Digite um valor em metros: '))
+n = float(input('Digite um valor em metros: '))
 print(f'{n} metros é {n*10**2} centímetros')
 print(f'{n} metros é {n*10**3} milímetros')
